Fixed issues with big CG log files.
The CGI helper script now returns chunks of 600KB for log files and
the number of pages available. API version bumped to 0.12.

The model that communicates with the helper CGI script accepts a seek
parameter from now on. Version bumped to 0.3.

The WCGP wrapper for the helper CGI script now accepts a seek parameter.

The Logs controller accepts seek parameter in the file method.